Fix a problem with disappearing cluster messages
The clustering method assumes it inserts the message
in the middle of the list. When two consecutive messages
have the same sender, the following message will be
inserted at the end of the list without updating the end pointer.

This is now corrected. We don't use the insert method
when the leading message was at the end of the list.